Dynamic vectorization into VIMA instructions (Master's degree version). Memory copy and vector sum patterns can be converted into VIMA instructions. Each converted instruction is locked inside ROB untill the conversion ends. In a successful conversion, the locked instructions are discarded. Otherwise, the correspondent VIMA result is discarded and they are executed in CPU as usual.

+++ branch_predictor/branch_predictor.hpp
@@ -18,7 +18,7 @@ class branch_predictor_t{
         // para acesso direto btb
         //===================================
         uint32_t index;
-        uint8_t way;
+        uint32_t way;
         //===================================
         //metodos para btb
         //===================================
